+++ addresses/index.js
@@ -1,3 +1,4 @@
 const decodeBase58Address = require('./decode_base58_address');
+const decodeBech32Address = require('./decode_bech32_address');
 
-module.exports = {decodeBase58Address};
+module.exports = {decodeBase58Address, decodeBech32Address};