+++ transactions/index.js
@@ -1,4 +1,5 @@
 const componentsOfTransaction = require('./components_of_transaction');
+const idForTransaction = require('./id_for_transaction');
 const noLocktimeIdForTransaction = require('./no_locktime_id_for_transaction');
 const parseTransaction = require('./parse_transaction');
 const queryTransactions = require('./query_transactions');
@@ -6,6 +7,7 @@ const unsignedTxFromPsbt = require('./unsigned_tx_from_psbt');
 
 module.exports = {
   componentsOfTransaction,
+  idForTransaction,
   noLocktimeIdForTransaction,
   parseTransaction,
   queryTransactions,